Everything is ok. Life is good. Free at last.

+++ backend/schemas.py
@@ -6,8 +6,8 @@ class Latest(BaseModel):
     user_id: int
     sleep_id: int
     ts: str
-    temperature: float
-    humidity: float
+    temperature: int
+    humidity: int
     heartrate: int
 
 class Efficiency(BaseModel):
@@ -25,8 +25,8 @@ class LogItem(BaseModel):
     user_id: int
     sleep_id: int
     ts: str
-    temperature: float
-    humidity: float
+    temperature: int
+    humidity: int
     heartrate: int
 
 class SessionInfo(BaseModel):
